spring logic handle home screen list product for guest

+++ src/main/java/spring/backend/m_bl5_g1_su25/OnlineShopping/ProductScreen/dto/response/ProductResponse.java
@@ -19,7 +19,7 @@ public class ProductResponse {
     String description;
     String imageUrl;
 
-    BigDecimal price;
+    Double price;
     Integer quantity;
     Integer sales_count = 0;
     @Enumerated(EnumType.STRING)
